+++ R/utils-upload.R
@@ -1,8 +1,8 @@
 #' Upload an image to imgur.com
 #'
-#' This function uses the \pkg{httr} package to upload a image to
+#' This function uses the \pkg{curl} package to upload a image to
 #' \url{https://imgur.com}, and parses the XML response to a list with
-#' \pkg{xml2} which contains information about the image in the Imgur website.
+#' \pkg{xml2}, which contains information about the image on Imgur.
 #'
 #' When the output format from \code{\link{knit}()} is HTML or Markdown, this
 #' function can be used to upload local image files to Imgur, e.g. set the
@@ -10,8 +10,10 @@
 #' document is completely self-contained, i.e. it does not need external image
 #' files any more, and it is ready to be published online.
 #' @param file Path to the image file to be uploaded.
-#' @param key Client ID for Imgur. By default, this uses a client ID registered
-#'   by Yihui Xie.
+#' @param key Client ID for Imgur. It can be set via either the global option
+#'   \code{knitr.imgur.key} or the environment variable \code{R_KNITR_IMGUR_KEY}
+#'   (see [xfun::env_option()]). If neither is set, this uses a client ID
+#'   registered by Yihui Xie.
 #' @return A character string of the link to the image; this string carries an
 #'   attribute named \code{XML} which is a list converted from the response XML
 #'   file; see Imgur API in the references.
@@ -31,20 +33,24 @@
 #' if (interactive()) browseURL(res)
 #'
 #' # to use your own key
-#' opts_knit$set(upload.fun = function(file) imgur_upload(file, key = 'your imgur key'))
+#' options(knitr.imgur.key = 'your imgur key')
 #' }
-imgur_upload = function(file, key = '9f3460e67f308f6') {
+imgur_upload = function(file, key = xfun::env_option('knitr.imgur.key', '9f3460e67f308f6')) {
   if (!is.character(key)) stop('The Imgur API Key must be a character string!')
-  resp = httr::POST(
-    "https://api.imgur.com/3/image.xml",
-    config = httr::add_headers(Authorization = paste("Client-ID", key)),
-    body = list(image = httr::upload_file(file))
+  h = curl::new_handle(httpheader = paste("Authorization: Client-ID", key))
+  curl::handle_setform(h, image = curl::form_file(file))
+  res = curl::curl_fetch_memory('https://api.imgur.com/3/image.xml', h)$content
+  if (loadable('xml2')) {
+    res = xml2::as_list(xml2::read_xml(res))
+    link = res[[1]]$link[[1]]
+  } else {
+    res = rawToChar(res)
+    link = xfun::grep_sub('.*<link>([^<]+)</link>.*', '\\1', res)
+  }
+  if (length(link) != 1) stop(
+    'Failed to upload ', file, sprintf(' (reason: %s)', if (is.character(res)) {
+      xfun::grep_sub('.*<error>([^<]+)</error>.*', '\\1', res)
+    } else res[[1]]$error[[1]])
   )
-  httr::stop_for_status(resp, "upload to imgur")
-  res = httr::content(resp, as = "raw")
-  res = if (length(res)) xml2::as_list(xml2::read_xml(res))
-  # Breaking change in xml2 1.2.0
-  if (packageVersion('xml2') >= '1.2.0') res <- res[[1L]]
-  if (is.null(res$link[[1]])) stop('failed to upload ', file)
-  structure(res$link[[1]], XML = res)
+  structure(link, XML = res)
 }